TextGenerationWithLocalLLM/requirements.txt
# LangChain Core
langchain

# Hugging face integration
langchain-huggingface
huggingface-hub

# Environment variable management
python-dotenv
